add contextmanager for prepending bin stuff to PATH in python process
