+++ conda_build/environ.py
@@ -17,7 +17,7 @@ from conda_build.os_utils import external
 from conda_build import source
 from conda_build import utils
 from conda_build.features import feature_list
-from conda_build.scripts import prepend_bin_path
+from conda_build.utils import prepend_bin_path
 
 
 def get_perl_ver(config):
